Okay, I've refactored the `s4p` TDR script into a more modular package for you.
Previously, the `tdr_plot_from-gemini-s4p.py` script was a single, monolithic file. This made it hard to reuse the core TDR analysis logic in other places.

I've now organized the code into a `s4p_tdr_analyzer` package with these distinct modules:
- `tdr.py`: For the core TDR calculations using the `s_params_to_tdr` function.
- `plotting.py`: For visualizing results with the `plot_tdr` function.
- `main.py`: An example of how to use the library, including test cases.
- `requirements.txt`: The list of package dependencies.

This new structure makes the code more maintainable and much easier to import and use as a library. As part of this cleanup, I've removed the original script.
