+++ s4p_tdr_analyzer/requirements.txt
@@ -1,0 +1,4 @@
+numpy
+matplotlib
+scipy
+scikit-rf
